reverse-engineering-assistant/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "reverse-engineering-assistant"
readme = "README.md"
version = "1.0.3"
authors = [
    {name="サイバーカイダ (cyberkaida)"},
]
description = "An AI assistant for reverse engineering tasks"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-openai",
    "prompt_toolkit",
    "sentence_transformers",
    "PyYAML",
    "pydantic",
    "rich",
    "grpcio",
    "protobuf",
]

[project.scripts]
revassistant = "reverse_engineering_assistant.assistant:main"
reva-serve = "reverse_engineering_assistant.assistant_api_server:main"
reva-server = "reverse_engineering_assistant.assistant_api_server:main"
reva-chat = "reverse_engineering_assistant.chat_client:main"